backend/requirements.txt
fastapi~=0.109.2
sqlmodel~=0.0.21
pydantic~=2.9.2
torch~=2.4.1
pyjwt~=2.9.0
numpy~=2.0.0
opencv-python~=4.10.0.84
pytesseract~=0.3.13
spacy~=3.8.2
pillow~=10.4.0
langdetect~=1.0.9
passlib~=1.7.4
pytest~=7.4.4
alembic~=1.13.2
sqlalchemy~=2.0.35
starlette~=0.36.3
emails~=0.6
jinja2~=3.1.4
tenacity~=8.5.0